Form submit for upload, list assignable categories.
Moved code from index.php into OOP structure. YouTubeBridge.php should
handle most of the YouTube API requests. YouTubeCategories is there
specifically to parse the categories xml provided by Google to see which
categories are assignable.

Currently, YouTubeCategories redownload the categories.cat url each time
it loads an instance of itself, this might cause problems if the server
it's on doesn't allow outside access or if the file does not exist
anymore. Probably best to read from a file and check for updates.