Add playlist insertion and creation.
The newly uploaded video can now be assigned to existing playlists or
have a new playlist be created for it.

Unfortunately, since all playlist manipulation has to occur with an
existing video, the playlist step can only come after the video has been
uploaded. A nicer user interface could pass along the playlist info step
after step, but that's annoying to do. So we'll just use a 3 step
process:
	1. Enter video metadata
	2. Upload video
	3. Modify playlists