Alphabetic sorting of playlists.
To be specific, it's a case insensitive sort, so that playlists shows up
in nice human order.

Playlists are stored in a array of arrays, and we want to sort by a
value of the sub-array, unfortunately, can't use built in php functions
for this. Could've built a custom comparison function, but just decided
to split it up into 1d arrays, sort one, and use that sorted one to
construct the return array.