Fix unable to get user playlist using login name.
A user's login name might not be the same as their Youtube username.
E.g.: The UBC youtube account is a legacy YouTube account that got
migrated over to a Google account. When logging in, you have to use a
different username. Once logged in, you see that the old YouTube
username has been retained for channels and feeds.

This problem is solved by passing in "default" as the username when
requesting the playlist feed. This tells Youtube to use the current
logged in user for the operation, letting them deal with the username
problem instead of us.